Rename main folder and hook
