+++ client/components/command-palette/index.tsx
@@ -8,7 +8,7 @@ import { Command, useCommandState } from 'cmdk';
 import { useEffect, useState, useRef, useMemo } from 'react';
 import { useSelector } from 'calypso/state';
 import getCurrentRoute from 'calypso/state/selectors/get-current-route';
-import { CommandCallBackParams, useCommandPallette } from './use-command-pallette';
+import { CommandCallBackParams, useCommandPallette } from './use-command-palette';
 
 import '@wordpress/commands/build-style/style.css';
