+++ client/components/command-palette/use-command-palette.tsx
